refactor: extract Form_Records_Post_Type from Form_Submissions
Move the otter_form_record post type, its read/unread statuses and the admin
capability grants into a dedicated Form_Records_Post_Type class registering its
own init/admin_init hooks. Update the test setup to grant caps via the new
class.
